+++ warp/stubs.py
@@ -750,13 +750,19 @@ def quat_to_axis_angle(quat: Quaternion[Float], axis: Vector[3, Float], angle: F
 
 @over
 def quat_from_matrix(mat: Matrix[3, 3, Float]) -> Quaternion[Float]:
-    """Construct a quaternion from a 3x3 matrix."""
+    """Construct a quaternion from a 3x3 matrix.
+
+    If the matrix is not a pure rotation, but for example includes scaling or skewing, the result is undefined.
+    """
     ...
 
 
 @over
 def quat_from_matrix(mat: Matrix[4, 4, Float]) -> Quaternion[Float]:
-    """Construct a quaternion from a 4x4 matrix."""
+    """Construct a quaternion from a 4x4 matrix.
+
+    If the top-left 3x3 block of the matrix is not a pure rotation, but for example includes scaling or skewing, the result is undefined.
+    """
     ...
 
 
@@ -1060,7 +1066,7 @@ def tile(x: Any) -> Tile:
 
     This function converts values computed using scalar kernel code to a tile representation for input into collective operations.
 
-    * If the input value is a scalar, then the resulting tile has ``shape=(1, block_dim)``
+    * If the input value is a scalar, then the resulting tile has ``shape=(block_dim,)``
     * If the input value is a vector, then the resulting tile has ``shape=(length(vector), block_dim)``
 
     :param x: A per-thread local value, e.g. scalar, vector, or matrix.
@@ -1153,13 +1159,12 @@ def tile_transpose(a: Tile) -> Tile:
 def tile_broadcast(a: Tile, shape: Tuple[int, ...]) -> Tile:
     """Broadcast a tile.
 
-    This function will attempt to broadcast the input tile ``a`` to the destination shape (m, n).
-
+    Broadcasts the input tile ``a`` to the destination shape.
     Broadcasting follows NumPy broadcast rules.
 
     :param a: Tile to broadcast
     :param shape: The shape to broadcast to
-    :returns: Tile with broadcast ``shape=(m, n)``
+    :returns: Tile with broadcast shape
     """
     ...